2.0.1次修改版中忘记修改的一下东西： 回调中获取当前播放时间，取消强转int类型，增加时间精度 AFSoundQueue.m的播放回调，播放完成后取消自动播放下一首

+++ Classes/AFSoundPlayback.m
@@ -67,8 +67,8 @@ NSString * const AFSoundPlaybackFinishedNotification = @"kAFSoundPlaybackFinishe
     }
     
     _feedbackTimer = [NSTimer scheduledTimerWithTimeInterval:updateRate block:^{
-        
-        _currentItem.timePlayed = (int)CMTimeGetSeconds(_player.currentTime);
+        // chenEdit 删除获取当前时间后，强制类型转换为int类型
+        _currentItem.timePlayed = CMTimeGetSeconds(_player.currentTime);
         
         if (block) {
